Make the cache before starting
