+++ server/server.go
@@ -51,7 +51,11 @@ type Server struct {
 }
 
 func NewServer(client *github.Client, user string) *Server {
-	s := &Server{client: client, user: user, ready: true}
+	s := &Server{
+		client:   client,
+		comments: map[string]*CommentCache{},
+		user:     user,
+		ready:    true}
 
 	s.authKey = os.Getenv("GHB_AUTH_TOKEN")
